requirements.txt
absl-py==2.0.0
alembic==1.13.1
amqp==5.2.0
aniso8601==9.0.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
docopt==0.6.2
Flask==3.0.0
Flask-Cors==4.0.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.47.0
gast==0.5.4
google-auth==2.26.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
idna==3.6
imagekitio==3.2.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
keras==2.15.0
kiwisolver==1.4.5
kombu==5.3.4
libclang==16.0.6
llvmlite==0.41.1
Mako==1.3.0
mariadb==1.0.11
Markdown==3.5.1
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.8.2
ml-dtypes==0.2.0
numba==0.58.1
numpy==1.26.3
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
pika==1.3.2
pillow==10.2.0
pipreqs==0.4.13
prompt-toolkit==3.0.43
protobuf==4.23.4
py==1.11.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
rabbitmq-pika-flask==1.2.35
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
retry==0.9.2
rpds-py==0.16.2
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
shap==0.44.0
six==1.16.0
slicer==0.0.7
SQLAlchemy==2.0.25
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
uvicorn==0.25.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.14.1
yarg==0.1.9
